chore: reorg and rename crates
