+++ Dockerfile
@@ -7,7 +7,7 @@ WORKDIR /app
 COPY Cargo.toml Cargo.lock ./
 COPY crates ./crates
 
-RUN cargo build --release -p attestation-service --bin attestation-service
+RUN cargo build --release -p attestation-api --bin attestation-api
 
 FROM debian:bookworm-slim
 
@@ -15,8 +15,8 @@ RUN apt-get update && \
     apt-get install -y --no-install-recommends libtss2-esys-3.0.2-0 libtss2-tctildr0 ca-certificates && \
     rm -rf /var/lib/apt/lists/*
 
-COPY --from=builder /app/target/release/attestation-service /attestation-service
+COPY --from=builder /app/target/release/attestation-api /attestation-api
 
 EXPOSE 8400
 
-ENTRYPOINT ["/attestation-service"]
+ENTRYPOINT ["/attestation-api"]
