+++ README.md
@@ -10,8 +10,8 @@ Rust workspace for TEE attestation libraries, tools, and services.
 | --- | --- | --- |
 | `attestation` | `crates/attestation` | Core TEE attestation evidence generation and verification library |
 | `attestation-cli` | `crates/attestation-cli` | CLI for generating and verifying attestation evidence |
-| `attestation-service` | `crates/attestation-service` | REST API service wrapping the attestation library |
-| `attestation-wasm-test` | `crates/attestation-wasm-test` | WASM verification harness |
+| `attestation-api` | `crates/attestation-api` | REST API service wrapping the attestation library |
+| `attestation-wasm` | `crates/attestation-wasm` | WASM verification harness |
 
 ## Common Commands
 
@@ -31,13 +31,13 @@ cargo build -p attestation-cli --release --features attest
 Build the REST service:
 
 ```bash
-cargo build -p attestation-service --release
+cargo build -p attestation-api --release
 docker build .
 ```
 
-The service image is published as `ghcr.io/lunal-dev/attestation-service`.
+The service image is published as `ghcr.io/lunal-dev/attestation-api`.
 
 ## Documentation
 
 - Core library: `crates/attestation/README.md`
-- REST service: `crates/attestation-service/README.md`
+- REST service: `crates/attestation-api/README.md`
