+++ crates/attestation-api/src/error.rs
