+++ crates/attestation-api/src/middleware/auth.rs
