+++ crates/attestation-wasm/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "attestation-wasm-test"
+name = "attestation-wasm"
 version = "0.1.0"
 edition = "2021"
